Update README.md
Advise users of option to use sudo instead of root.
Docker version 0.5.2 daemon binds to Unix socket, which, by default, is owned by user root.  For this reason the docker daemon must always run as the root user.  It is possible to add a user to a docker group, in doing so that user would gain root privileges and not need to use 'sudo'.